Updated readme, added more organization to the components directory for each feature, and added the popup feature
Moved all of the presentation/container components to a new directory under the components directory
and added the componets for a popup dialogue. Updated the readme to use better commands, list the new
feature/pattern/component thing, and stipulated that a minimum version of nodejs is required to run
the project.