Server side render now takes data from POST request made to it
When rendering on the server side, there is a need to be able to pass into react
components to show the initial state (using the data) of that component.
Previously, the server side code was not able to process data passed into a
request's body. Now it can and it will pass the data to the rendered component.